[#49]Refactor: Facade 패턴에서 MapperClass로 변경
