+++ src/main/java/com/project/bumawiki/domain/docs/service/DocsCheckYouLikeThisService.java
@@ -1,7 +1,7 @@
 package com.project.bumawiki.domain.docs.service;
 
 import com.project.bumawiki.domain.docs.domain.Docs;
-import com.project.bumawiki.domain.docs.facade.DocsFacade;
+import com.project.bumawiki.domain.docs.domain.repository.DocsRepositoryMapper;
 import com.project.bumawiki.domain.user.UserFacade;
 import com.project.bumawiki.domain.user.domain.User;
 import com.project.bumawiki.global.error.exception.ErrorCode;
@@ -13,12 +13,12 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class DocsCheckYouLikeThisService {
-    private final DocsFacade docsFacade;
+    private final DocsRepositoryMapper docsRepositoryMapper;
     private final UserFacade userFacade;
 
     public boolean checkUserLikeThisDocs(final Long docsId) {
         User currentUser = userFacade.getCurrentUser();
-        Docs docs = docsFacade.findById(docsId, ErrorCode.DOCS_NOT_FOUND);
+        Docs docs = docsRepositoryMapper.findById(docsId, ErrorCode.DOCS_NOT_FOUND);
         User foundUser = userFacade.getUserByEmail(currentUser.getEmail());
 
         return docs.doesUserLike(foundUser);
