+++ src/main/java/com/project/bumawiki/domain/thumbsUp/service/ThumbsUpManipulateService.java
@@ -1,7 +1,7 @@
 package com.project.bumawiki.domain.thumbsUp.service;
 
 import com.project.bumawiki.domain.docs.domain.Docs;
-import com.project.bumawiki.domain.docs.facade.DocsFacade;
+import com.project.bumawiki.domain.docs.domain.repository.DocsRepositoryMapper;
 import com.project.bumawiki.domain.thumbsUp.domain.ThumbsUp;
 import com.project.bumawiki.domain.thumbsUp.presentation.dto.ThumbsUpRequestDto;
 import com.project.bumawiki.domain.user.domain.User;
@@ -16,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class ThumbsUpManipulateService {
-    private final DocsFacade docsFacade;
+    private final DocsRepositoryMapper docsRepositoryMapper;
     private final UserRepository userRepository;
 
     @Transactional
@@ -59,7 +59,7 @@ public class ThumbsUpManipulateService {
 
     //Docs, User 가져오기
     private Docs getDocs(ThumbsUpRequestDto likeRequestDto) {
-        return docsFacade.findById(
+        return docsRepositoryMapper.findById(
                 likeRequestDto.getDocsId(),
                 ErrorCode.NO_DOCS_YOU_THUMBS_UP
         );
